+++ audiomate/corpus/io/__init__.py
@@ -20,6 +20,7 @@ from .urbansound import Urbansound8kReader  # noqa: F401
 from .timit import TimitReader  # noqa: F401
 from .free_spoken_digits import FreeSpokenDigitDownloader, FreeSpokenDigitReader  # noqa: F401
 from .tatoeba import TatoebaDownloader, TatoebaReader  # noqa: F401
+from .common_voice import CommonVoiceDownloader, CommonVoiceReader  # noqa: F401
 
 __downloaders = {}
 for cls in CorpusDownloader.__subclasses__():